OPG-530: Move the packaging scripts and their inputs under scripts/
Consolidates everything that builds a distributable into one place, one
directory per artifact type:

  scripts/rpm/  make-rpm.js, spec.mustache   (from makerpm.js, src/rpm/)
  scripts/deb/  make-deb.js, debian/         (from makedeb.js, src/debian/)
  scripts/zip/  make-zip.js                  (from makezip.js)

The templates in particular do not belong under src/: webpack sweeps src/
into dist with broad copy globs and `npm run sign` then attests everything
in dist into a signed MANIFEST.txt, so a packaging input with a matching
extension would ship in the plugin and break the signature. They were safe
only because the globs happened not to match their extensions — the same
hazard as the dist/test bug fixed in OPG-527.

Adds `npm run package:rpm|deb|zip` so .circleci/config.yml invokes the
scripts by name rather than hardcoding their paths.

Beyond the paths, the three entry points now resolve dist/ and artifacts/
from __dirname rather than process.cwd(). That was only safe while they sat
at the repository root; they no longer do.

Verified: full jest suite, typecheck, lint and `npm run build` all as
before, plus a real end-to-end build of all three artifacts (deb via the
CI debian image). Payloads and deb control metadata are unchanged, and the
rendered-spec and debian/control golden snapshots still match.
