+++ scripts/deb/make-deb.js
@@ -10,12 +10,12 @@ const rimraf = require('rimraf');
 const which = require('which');
 const program = require('commander');
 
-const { recursiveCopyFilter } = require('./scripts/distContents');
-const { resolveMaintainer } = require('./scripts/deb/maintainer');
-const { renderChangelog, renderControl } = require('./scripts/deb/metadata');
-const { resolveVersionAndRelease } = require('./scripts/packageVersion');
-const pkginfo = require('./package.json');
-const plugininfo = require('./src/plugin.json');
+const { recursiveCopyFilter } = require('../distContents');
+const { resolveMaintainer } = require('./maintainer');
+const { renderChangelog, renderControl, PROJECT_DIR } = require('./metadata');
+const { resolveVersionAndRelease } = require('../packageVersion');
+const pkginfo = require('../../package.json');
+const plugininfo = require('../../src/plugin.json');
 
 try {
   which.sync('dpkg-buildpackage');
@@ -43,9 +43,11 @@ pkginfo.release = release;
 // generated from one value. DEBFULLNAME/DEBEMAIL override it.
 const maintainer = resolveMaintainer();
 
+// Resolved from this file rather than from process.cwd(): the script no longer sits
+// at the project root, so the two are only the same by convention.
 const pkgid   = plugininfo.id;
-const workdir = path.join(process.cwd(), 'artifacts', pkgid);
-const distdir = path.join(process.cwd(), 'dist');
+const workdir = path.join(PROJECT_DIR, 'artifacts', pkgid);
+const distdir = path.join(PROJECT_DIR, 'dist');
 
 rimraf.sync(workdir);
 fs.mkdirsSync(workdir);
@@ -65,7 +67,7 @@ return copy(distdir, workdir, {
   fs.mkdirsSync(debian);
 
   // The templates are rendered below; they must not be copied through as-is.
-  return copy(path.join(process.cwd(), 'src', 'debian'), debian, {
+  return copy(path.join(__dirname, 'debian'), debian, {
     filter: ['**/*', '!**/*.mustache']
   }).then((_copyResults) => {
     console.log('debian/ directory copied');
@@ -80,7 +82,7 @@ return copy(distdir, workdir, {
     console.log('* running dpkg-buildpackage');
     // -us -uc: do not sign the .dsc/.changes here. dpkg-buildpackage would sign as the
     // changelog's maintainer identity, which is not a key we hold; CI signs the built
-    // .deb separately with the OpenNMS release key. Before makedeb.js checked the exit
+    // .deb separately with the OpenNMS release key. Before make-deb.js checked the exit
     // status, that signing failure (status 25) was discarded and the build looked green.
     const ret = spawn('dpkg-buildpackage', ['-us', '-uc'], {
       cwd: workdir,
